+++ app/routes/data-api/project/recordings.js
@@ -8,7 +8,7 @@ const stream = require('stream');
 const moment = require('moment');
 const dayInMs = 24 * 60 * 60 * 1000;
 var config = require('../../../config');
-var mime = require('mime');
+const mime = require('mime');
 
 let s3, s3RFCx;
 let cachedData = {
@@ -218,7 +218,11 @@ processFiltersData = async function(req, res, next) {
     }
     // Combine grouped detections report.
     if (projectionFilter && projectionFilter.grouped && projectionFilter.validation && !projectionFilter.species) {
-        return model.recordings.exportRecordingData(projectionFilter, filters).then(async function(results) {
+        let sites
+        model.projects.getProjectSites(filters.project_id).then(function(rows) {
+            sites = rows.map(s=>s.name)
+        })
+        return model.recordings.groupedDetections(projectionFilter, filters).then(async function(results) {
             let gKey = projectionFilter.grouped;
             let fields = [];
             results.forEach(result => {
@@ -238,6 +242,18 @@ processFiltersData = async function(req, res, next) {
                     data[s][f] += r[f] === '---' || r[f] === undefined ? 0 : +r[f];
                 })
             });
+            let addSites = sites.filter(s => !Object.keys(data).includes(s))
+            if (addSites && addSites.length && gKey === 'site') {
+                addSites.forEach(s => {
+                    if (!data[s]) {
+                        data[s] = {};
+                        data[s][gKey] = s;
+                    };
+                    fields.forEach((f) => {
+                        if (data[s][f] === undefined) { data[s][f] = 0 };
+                    })
+                })
+            }
             fields.unshift(gKey);
             let datastream = new stream.Readable({objectMode: true});
                 let streamArray = Object.values(data);
@@ -319,6 +335,8 @@ async function downloadRecordingById(req, res, next) {
     recording[0].name = namePartials[namePartials.length - 1];
     let legacy = recording[0].uri.startsWith('project_');
     res.set({'Content-Disposition' : `attachment; filename=${recording[0].name}`});
+    let mimetype = mime.getType(recording[0].name);
+    res.setHeader('Content-type', mimetype);
     await getRecordingFromS3(config(legacy? 'aws' : 'aws-rfcx').bucketName, legacy, recording[0].uri, res);
 }
 
@@ -429,6 +447,11 @@ router.param('oneRecUrl', function(req, res, next, recording_url){
         if(!recordings.length){
             return res.status(404).json({ error: "recording not found"});
         }
+        let recExt;
+        if (recordings[0].file) {
+            recExt = path.extname(recordings[0].file);
+            recordings[0].ext = recExt;
+        }
         req.recording = recordings[0];
         return next();
     });
@@ -473,7 +496,11 @@ router.get('/:get/:oneRecUrl?', function(req, res, next) {
     switch(get){
         case 'info'  :
             var url_comps = /(.*)\/([^/]+)\/([^/]+)/.exec(req.originalUrl);
-            recording.audioUrl = url_comps[1] + "/audio/" + recording.id;
+            let recExt;
+            if (recording.file) {
+                recExt = path.extname(recording.file);
+            }
+            recording.audioUrl = url_comps[1] + "/audio/" + recording.id + (recExt ? recExt : '');
             recording.imageUrl = url_comps[1] + "/image/" + recording.id;
             model.recordings.fetchValidations(recording, function(err, validations){
                 if(err) return next(err);