+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Arbimon Release Notes
 
-- CE-1414 The issue with updating projects is fixed
+## v3.0.39 - October 07, 2021
 
-## v3.0.38 - September XX, 2021
+Resolved issues:
+
+- The issue with getting all species grouped detections is fixed
+- CE-1398 Regex logic is updated for audio path
+- CE-1398 The inability to hear recordings is fixed in the Visualizer page via Safari browser
+
+## v3.0.38 - October 01, 2021
 
 New feature:
 
@@ -10,12 +16,12 @@ New feature:
 
 Resolved issues:
 
-- CE-1398 The inability to hear recordings is fixed in the Visualizer page via Safari browser
 - The label is wrapping to the next strings for the final box on the spectrogram
 - Fixed the issue with changing the position of a recording
 - CE-1371 The migration popup is deleted
 - CE-1342 ROI template issue is fixed
 - The Longitude issue is fixed
+- CE-1414 The issue with updating projects is fixed
 
 ## v3.0.37 - September 16, 2021